requirements.txt
# Python
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic[email]==2.5.3
email-validator==2.1.0
jinja2==3.1.3
python-dotenv==1.0.0
